nn/requirements.txt
numpy==1.23.4
Pillow==9.3.0
albumentations==1.3.0
scikit-learn==1.1.3
-f https://download.pytorch.org/whl/cu117
torch
torchvision
torchaudio
pytorch-metric-learning==1.6.3
transformers==4.24.0
pandas==1.5.1
scipy==1.9.3
tqdm==4.64.1